+++ src/SoilTracers.jl
@@ -1,0 +1,7 @@
+__precompile__()
+
+module SoilTracers
+
+include("consts.jl")
+
+end # module