+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+* Unit tests for `SoilTracers.Constants`.
 * Functions that calculate water properties (for aqueous reactions).
 
 ## 0.0.1 - 2018-05-04